feat: add related issues, pr labels
