+++ public/src/app/pull-request/pull-request.component.css
@@ -145,8 +145,33 @@
  .title {
    display: block;
    font-weight: bold;
+   margin-left: 4px;
  }
 
  .status {
    display: block;
+   margin-left: 4px;
  }
+
+ .issue-label {
+  margin-top: 2px;
+  margin-left: 3px;
+  border-radius: 10px;
+  padding: 2px 6px;
+  font-weight: 500;
+  font-size: 8pt;
+  word-wrap: break-word;
+  word-break: normal;
+  white-space: normal;
+  display: inline-block;
+}
+
+.related-issue {
+  padding: 0 4px;
+  margin: 0 4px;
+}
+
+.related-issue:hover {
+  background: #c0ffff;
+  border-radius: 3px 3px 0 0;
+}
