+++ public/src/app/pull-request/pull-request.component.html
@@ -7,11 +7,14 @@
 </span>
 <div class="pull-request-detail">
   <span class="branches">
-    <span class="branch">{{pull.pull.node.baseRefName}}</span> &larr; <span class="branch">{{pull.pull.node.headRefName}}</span> 
+    <span class="branch">{{pull.pull.node.baseRefName}}</span> &larr; <span class="branch">{{pull.pull.node.headRefName}}</span>
+    <span class="related-issue" *ngFor="let issue of pull.pull.node.timelineItems?.nodes">
+      <a href="https://github.com/keymanapp/keyman/issues/{{issue.subject.number}}" target="_blank">🔗 #{{issue.subject.number}}</a>
+    </span>
     <app-clipboard [text]="pull.pull.node.headRefName" title="Copy head branch name to clipboard"></app-clipboard>
   </span>
-  
   <span class="title">{{pull.pull.node.title}}</span>
   <span class="status">{{pull.state ? pull.state.context + ': ' + pull.state.description : ''}}</span>
+  <span class="labels"><span class="issue-label" *ngFor="let label of pull.pull.node.labels.edges" [style]="labelColor(label.node)">{{labelName(label.node.name)}}</span></span>
 </div>
 </span>
